Move ensure terminal admission stash to ScoresExportContext.
Keeps inference-specific ensure side effects out of generic export query context and restores catalog import isolation.
